fix telemetry tests
