+++ pkg/telemetry/events_test.go
@@ -1,8 +1,9 @@
-package telemetrytest
+package telemetry_test
 
 import (
 	"context"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 
@@ -34,6 +35,7 @@ func Test_OnParticipantJoin_EventIsSent(t *testing.T) {
 
 	// do
 	fixture.sut.ParticipantJoined(context.Background(), room, participantInfo, clientInfo, clientMeta)
+	time.Sleep(time.Millisecond * 500)
 
 	// test
 	require.Equal(t, 1, fixture.analytics.SendEventCallCount())
@@ -68,6 +70,7 @@ func Test_OnParticipantLeft_EventIsSent(t *testing.T) {
 
 	// do
 	fixture.sut.ParticipantLeft(context.Background(), room, participantInfo)
+	time.Sleep(time.Millisecond * 500)
 
 	// test
 	require.Equal(t, 1, fixture.analytics.SendEventCallCount())
@@ -102,6 +105,7 @@ func Test_OnTrackUpdate_EventIsSent(t *testing.T) {
 
 	// do
 	fixture.sut.TrackPublishedUpdate(context.Background(), livekit.ParticipantID(partID), trackInfo)
+	time.Sleep(time.Millisecond * 500)
 
 	// test
 	require.Equal(t, 1, fixture.analytics.SendEventCallCount())
@@ -142,6 +146,7 @@ func Test_OnParticipantActive_EventIsSent(t *testing.T) {
 
 	// do
 	fixture.sut.ParticipantJoined(context.Background(), room, participantInfo, clientInfo, clientMeta)
+	time.Sleep(time.Millisecond * 500)
 
 	// test
 	require.Equal(t, 1, fixture.analytics.SendEventCallCount())
@@ -154,6 +159,7 @@ func Test_OnParticipantActive_EventIsSent(t *testing.T) {
 	}
 
 	fixture.sut.ParticipantActive(context.Background(), room, participantInfo, clientMetaConnect)
+	time.Sleep(time.Millisecond * 500)
 
 	require.Equal(t, 2, fixture.analytics.SendEventCallCount())
 	_, eventActive := fixture.analytics.SendEventArgsForCall(1)
@@ -192,16 +198,16 @@ func Test_OnTrackSubscribed_EventIsSent(t *testing.T) {
 
 	// do
 	fixture.sut.ParticipantJoined(context.Background(), room, participantInfo, clientInfo, clientMeta)
+	time.Sleep(time.Millisecond * 500)
 
 	// test
 	require.Equal(t, 1, fixture.analytics.SendEventCallCount())
 	_, event := fixture.analytics.SendEventArgsForCall(0)
 	require.Equal(t, room, event.Room)
 
-	// test
 	// do
-
 	fixture.sut.TrackSubscribed(context.Background(), livekit.ParticipantID(partSID), trackInfo, publisherInfo)
+	time.Sleep(time.Millisecond * 500)
 
 	require.Equal(t, 2, fixture.analytics.SendEventCallCount())
 	_, eventTrackSubscribed := fixture.analytics.SendEventArgsForCall(1)
